Move `CheckClusterNodes` to shoot care
This functionality is very specific to shoot clusters and should therefore reside in the shoot care related package.
