+++ pkg/operation/care/health_test.go
@@ -21,13 +21,16 @@ import (
 
 	"github.com/Masterminds/semver"
 	druidv1alpha1 "github.com/gardener/etcd-druid/api/v1alpha1"
+	machinev1alpha1 "github.com/gardener/machine-controller-manager/pkg/apis/machine/v1alpha1"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	. "github.com/onsi/gomega/gstruct"
 	"github.com/onsi/gomega/types"
+	"go.uber.org/mock/gomock"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
 	testclock "k8s.io/utils/clock/testing"
 	"k8s.io/utils/pointer"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -37,6 +40,8 @@ import (
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
 	"github.com/gardener/gardener/pkg/client/kubernetes"
+	"github.com/gardener/gardener/pkg/component/extensions/operatingsystemconfig/executor"
+	mockclient "github.com/gardener/gardener/pkg/mock/controller-runtime/client"
 	. "github.com/gardener/gardener/pkg/operation/care"
 	healthchecker "github.com/gardener/gardener/pkg/utils/kubernetes/health/checker"
 	. "github.com/gardener/gardener/pkg/utils/test/matchers"
@@ -172,12 +177,10 @@ var _ = Describe("health check", func() {
 		fakeClient = fakeclient.NewClientBuilder().WithScheme(kubernetes.SeedScheme).Build()
 		fakeClock = testclock.NewFakeClock(time.Now())
 		condition = gardencorev1beta1.Condition{
-			Type:               gardencorev1beta1.ConditionType("test"),
+			Type:               "test",
 			LastTransitionTime: metav1.Time{Time: fakeClock.Now()},
 		}
 	})
-
-	var ()
 
 	DescribeTable("#CheckShootontrolPlane",
 		func(shoot *gardencorev1beta1.Shoot, deployments []*appsv1.Deployment, etcds []*druidv1alpha1.Etcd, workers []*extensionsv1alpha1.Worker, conditionMatcher types.GomegaMatcher) {
@@ -714,6 +717,417 @@ var _ = Describe("health check", func() {
 			nil,
 			ConsistOf(beConditionWithStatus(gardencorev1beta1.ConditionFalse))),
 	)
+
+	Describe("#CheckClusterNodes", func() {
+		var (
+			ctrl *gomock.Controller
+			c    *mockclient.MockClient
+
+			workerPoolName1            = "cpu-worker-1"
+			workerPoolName2            = "cpu-worker-2"
+			cloudConfigSecretChecksum1 = "foo"
+			cloudConfigSecretChecksum2 = "foo"
+			nodeName                   = "node1"
+			cloudConfigSecretMeta      = map[string]metav1.ObjectMeta{
+				workerPoolName1: {
+					Name:        "cloud-config-cpu-worker-1-c63c0",
+					Labels:      map[string]string{"worker.gardener.cloud/pool": workerPoolName1},
+					Annotations: map[string]string{"checksum/data-script": cloudConfigSecretChecksum1},
+				},
+				workerPoolName2: {
+					Name:        "bar",
+					Labels:      map[string]string{"worker.gardener.cloud/pool": workerPoolName2},
+					Annotations: map[string]string{"checksum/data-script": cloudConfigSecretChecksum2},
+				},
+			}
+		)
+
+		BeforeEach(func() {
+			ctrl = gomock.NewController(GinkgoT())
+			c = mockclient.NewMockClient(ctrl)
+		})
+
+		AfterEach(func() {
+			ctrl.Finish()
+		})
+
+		DescribeTable("#CheckClusterNodes",
+			func(nodes []corev1.Node, workerPools []gardencorev1beta1.Worker, cloudConfigSecretMeta map[string]metav1.ObjectMeta, conditionMatcher types.GomegaMatcher) {
+				c.EXPECT().List(ctx, gomock.AssignableToTypeOf(&corev1.NodeList{})).DoAndReturn(func(_ context.Context, list *corev1.NodeList, _ ...client.ListOption) error {
+					*list = corev1.NodeList{Items: nodes}
+					return nil
+				})
+				cloudConfigSecretListOptions := []client.ListOption{
+					client.InNamespace(metav1.NamespaceSystem),
+					client.MatchingLabels{"gardener.cloud/role": "cloud-config"},
+				}
+				c.EXPECT().List(ctx, gomock.AssignableToTypeOf(&corev1.SecretList{}), cloudConfigSecretListOptions).DoAndReturn(func(_ context.Context, list *corev1.SecretList, _ ...client.ListOption) error {
+					*list = corev1.SecretList{}
+					for _, meta := range cloudConfigSecretMeta {
+						list.Items = append(list.Items, corev1.Secret{
+							ObjectMeta: meta,
+						})
+					}
+					return nil
+				})
+
+				checker := healthchecker.NewHealthChecker(fakeClient, fakeClock, map[gardencorev1beta1.ConditionType]time.Duration{}, nil, nil, nil, kubernetesVersion)
+
+				exitCondition, err := CheckClusterNodes(ctx, c, fakeClient, checker, kubernetesVersion, seedNamespace, workerPools, condition)
+				Expect(err).NotTo(HaveOccurred())
+				Expect(exitCondition).To(conditionMatcher)
+			},
+			Entry("all healthy",
+				[]corev1.Node{
+					newNode(nodeName, true, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, nil, kubernetesVersion.Original()),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+				},
+				cloudConfigSecretMeta,
+				BeNil()),
+			Entry("node not healthy",
+				[]corev1.Node{
+					newNode(nodeName, false, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, nil, kubernetesVersion.Original()),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+				},
+				cloudConfigSecretMeta,
+				PointTo(beConditionWithStatusAndMsg(gardencorev1beta1.ConditionFalse, "NodeUnhealthy", fmt.Sprintf("Node %q in worker group %q is unhealthy", nodeName, workerPoolName1)))),
+			Entry("node not healthy with error codes",
+				[]corev1.Node{
+					{
+						ObjectMeta: metav1.ObjectMeta{
+							Name:   nodeName,
+							Labels: labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"},
+						},
+						Status: corev1.NodeStatus{
+							Conditions: []corev1.NodeCondition{
+								{
+									Type:   corev1.NodeReady,
+									Status: corev1.ConditionTrue,
+								},
+								{
+									Type:   corev1.NodeDiskPressure,
+									Status: corev1.ConditionTrue,
+									Reason: "KubeletHasDiskPressure",
+								},
+							},
+						},
+					},
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+				},
+				cloudConfigSecretMeta,
+				PointTo(beConditionWithStatusAndCodes(gardencorev1beta1.ConditionFalse, gardencorev1beta1.ErrorConfigurationProblem))),
+			Entry("not enough nodes in worker pool",
+				[]corev1.Node{
+					newNode(nodeName, true, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, nil, kubernetesVersion.Original()),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+					{
+						Name:    workerPoolName2,
+						Maximum: 2,
+						Minimum: 1,
+					},
+				},
+				cloudConfigSecretMeta,
+				PointTo(beConditionWithStatusAndMsg(gardencorev1beta1.ConditionFalse, "MissingNodes", fmt.Sprintf("Not enough worker nodes registered in worker pool %q to meet minimum desired machine count. (%d/%d).", workerPoolName2, 0, 1)))),
+			Entry("not enough nodes in worker pool",
+				[]corev1.Node{
+					newNode(nodeName, true, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, nil, kubernetesVersion.Original()),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+					{
+						Name:    workerPoolName2,
+						Maximum: 2,
+						Minimum: 1,
+					},
+				},
+				cloudConfigSecretMeta,
+				PointTo(beConditionWithStatusAndMsg(gardencorev1beta1.ConditionFalse, "MissingNodes", fmt.Sprintf("Not enough worker nodes registered in worker pool %q to meet minimum desired machine count. (%d/%d).", workerPoolName2, 0, 1)))),
+			Entry("too old Kubernetes patch version",
+				[]corev1.Node{
+					newNode(nodeName, true, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, nil, "v1.23.2"),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+				},
+				cloudConfigSecretMeta,
+				PointTo(beConditionWithStatusAndMsg(gardencorev1beta1.ConditionFalse, "KubeletVersionMismatch", fmt.Sprintf("The kubelet version for node %q (v1.23.2) does not match the desired Kubernetes version (v%s)", nodeName, kubernetesVersion.Original())))),
+			Entry("same Kubernetes patch version",
+				[]corev1.Node{
+					newNode(nodeName, true, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, nil, "v1.23.3"),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+				},
+				cloudConfigSecretMeta,
+				BeNil()),
+			Entry("too old Kubernetes patch version with pool version overwrite",
+				[]corev1.Node{
+					newNode(nodeName, true, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, nil, "v1.22.2"),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+						Kubernetes: &gardencorev1beta1.WorkerKubernetes{
+							Version: pointer.String("1.22.3"),
+						},
+					},
+				},
+				cloudConfigSecretMeta,
+				PointTo(beConditionWithStatusAndMsg(gardencorev1beta1.ConditionFalse, "KubeletVersionMismatch", fmt.Sprintf("The kubelet version for node %q (v1.22.2) does not match the desired Kubernetes version (v1.22.3)", nodeName)))),
+			Entry("different Kubernetes minor version (all healthy)",
+				[]corev1.Node{
+					newNode(nodeName, true, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, nil, "v1.22.2"),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+				},
+				cloudConfigSecretMeta,
+				BeNil()),
+			Entry("missing cloud-config secret checksum for a worker pool",
+				[]corev1.Node{
+					newNode(nodeName, true, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, nil, "v1.22.2"),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+				},
+				nil,
+				PointTo(beConditionWithStatusAndMsg(gardencorev1beta1.ConditionFalse, "CloudConfigOutdated", fmt.Sprintf("missing cloud config secret metadata for worker pool %q", workerPoolName1)))),
+			Entry("no cloud-config node checksum for a worker pool",
+				[]corev1.Node{
+					newNode(nodeName, true, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, nil, "v1.22.2"),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+				},
+				cloudConfigSecretMeta,
+				BeNil()),
+			Entry("outdated cloud-config secret checksum for a worker pool",
+				[]corev1.Node{
+					newNode(nodeName, true, labels.Set{"worker.gardener.cloud/pool": workerPoolName1, "worker.gardener.cloud/kubernetes-version": "1.24.0"}, map[string]string{executor.AnnotationKeyChecksum: "outdated"}, "v1.22.2"),
+				},
+				[]gardencorev1beta1.Worker{
+					{
+						Name:    workerPoolName1,
+						Maximum: 10,
+						Minimum: 1,
+					},
+				},
+				map[string]metav1.ObjectMeta{
+					workerPoolName1: {
+						Name:        "cloud-config-cpu-worker-1-c63c0",
+						Annotations: map[string]string{"checksum/data-script": cloudConfigSecretChecksum1},
+						Labels:      map[string]string{"worker.gardener.cloud/pool": workerPoolName1},
+					},
+				},
+				PointTo(beConditionWithStatusAndMsg(gardencorev1beta1.ConditionFalse, "CloudConfigOutdated", fmt.Sprintf("the last successfully applied cloud config on node %q is outdated", nodeName)))),
+		)
+	})
+
+	Describe("#CheckNodesScalingUp", func() {
+		It("should return true if number of ready nodes equal number of desired machines", func() {
+			Expect(CheckNodesScalingUp(nil, 1, 1)).To(Succeed())
+		})
+
+		It("should return an error if not enough machine objects as desired were created", func() {
+			Expect(CheckNodesScalingUp(&machinev1alpha1.MachineList{}, 0, 1)).To(MatchError(ContainSubstring("not enough machine objects created yet")))
+		})
+
+		It("should return an error when detecting erroneous machines", func() {
+			machineList := &machinev1alpha1.MachineList{
+				Items: []machinev1alpha1.Machine{
+					{
+						Status: machinev1alpha1.MachineStatus{
+							CurrentStatus: machinev1alpha1.CurrentStatus{Phase: machinev1alpha1.MachineUnknown},
+						},
+					},
+				},
+			}
+
+			Expect(CheckNodesScalingUp(machineList, 0, 1)).To(MatchError(ContainSubstring("is erroneous")))
+		})
+
+		It("should return an error when not enough ready nodes are registered", func() {
+			machineList := &machinev1alpha1.MachineList{
+				Items: []machinev1alpha1.Machine{
+					{
+						Status: machinev1alpha1.MachineStatus{
+							CurrentStatus: machinev1alpha1.CurrentStatus{Phase: machinev1alpha1.MachineRunning},
+						},
+					},
+				},
+			}
+
+			Expect(CheckNodesScalingUp(machineList, 0, 1)).To(MatchError(ContainSubstring("not enough ready worker nodes registered in the cluster")))
+		})
+
+		It("should return progressing when detecting a regular scale up (pending status)", func() {
+			machineList := &machinev1alpha1.MachineList{
+				Items: []machinev1alpha1.Machine{
+					{
+						Status: machinev1alpha1.MachineStatus{
+							CurrentStatus: machinev1alpha1.CurrentStatus{Phase: machinev1alpha1.MachinePending},
+						},
+					},
+				},
+			}
+
+			Expect(CheckNodesScalingUp(machineList, 0, 1)).To(MatchError(ContainSubstring("is provisioning and should join the cluster soon")))
+		})
+
+		It("should return progressing when detecting a regular scale up (no status)", func() {
+			machineList := &machinev1alpha1.MachineList{
+				Items: []machinev1alpha1.Machine{
+					{},
+				},
+			}
+
+			Expect(CheckNodesScalingUp(machineList, 0, 1)).To(MatchError(ContainSubstring("is provisioning and should join the cluster soon")))
+		})
+	})
+
+	Describe("#CheckNodesScalingDown", func() {
+		It("should return true if number of registered nodes equal number of desired machines", func() {
+			Expect(CheckNodesScalingDown(nil, nil, 1, 1)).To(Succeed())
+		})
+
+		It("should return an error if the machine for a cordoned node is not found", func() {
+			nodeList := &corev1.NodeList{
+				Items: []corev1.Node{
+					{Spec: corev1.NodeSpec{Unschedulable: true}},
+				},
+			}
+
+			Expect(CheckNodesScalingDown(&machinev1alpha1.MachineList{}, nodeList, 2, 1)).To(MatchError(ContainSubstring("machine object for cordoned node \"\" not found")))
+		})
+
+		It("should return an error if the machine for a cordoned node is not deleted", func() {
+			var (
+				nodeName = "foo"
+
+				machineList = &machinev1alpha1.MachineList{
+					Items: []machinev1alpha1.Machine{
+						{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{"node": nodeName}}},
+					},
+				}
+				nodeList = &corev1.NodeList{
+					Items: []corev1.Node{
+						{
+							ObjectMeta: metav1.ObjectMeta{Name: nodeName},
+							Spec:       corev1.NodeSpec{Unschedulable: true},
+						},
+					},
+				}
+			)
+
+			Expect(CheckNodesScalingDown(machineList, nodeList, 2, 1)).To(MatchError(ContainSubstring("found but corresponding machine object does not have a deletion timestamp")))
+		})
+
+		It("should return an error if there are more nodes then machines", func() {
+			Expect(CheckNodesScalingDown(&machinev1alpha1.MachineList{}, &corev1.NodeList{Items: []corev1.Node{{}}}, 2, 1)).To(MatchError(ContainSubstring("too many worker nodes are registered. Exceeding maximum desired machine count")))
+		})
+
+		It("should return progressing for a regular scale down", func() {
+			var (
+				nodeName          = "foo"
+				deletionTimestamp = metav1.Now()
+
+				machineList = &machinev1alpha1.MachineList{
+					Items: []machinev1alpha1.Machine{
+						{ObjectMeta: metav1.ObjectMeta{DeletionTimestamp: &deletionTimestamp, Labels: map[string]string{"node": nodeName}}},
+					},
+				}
+				nodeList = &corev1.NodeList{
+					Items: []corev1.Node{
+						{
+							ObjectMeta: metav1.ObjectMeta{Name: nodeName},
+							Spec:       corev1.NodeSpec{Unschedulable: true},
+						},
+					},
+				}
+			)
+
+			Expect(CheckNodesScalingDown(machineList, nodeList, 2, 1)).To(MatchError(ContainSubstring("is waiting to be completely drained from pods")))
+		})
+
+		It("should ignore node not managed by MCM and return progressing for a regular scale down", func() {
+			var (
+				nodeName          = "foo"
+				deletionTimestamp = metav1.Now()
+
+				machineList = &machinev1alpha1.MachineList{
+					Items: []machinev1alpha1.Machine{
+						{ObjectMeta: metav1.ObjectMeta{DeletionTimestamp: &deletionTimestamp, Labels: map[string]string{"node": nodeName}}},
+					},
+				}
+				nodeList = &corev1.NodeList{
+					Items: []corev1.Node{
+						{
+							ObjectMeta: metav1.ObjectMeta{Name: nodeName},
+							Spec:       corev1.NodeSpec{Unschedulable: true},
+						},
+						{
+							ObjectMeta: metav1.ObjectMeta{
+								Name:        "bar",
+								Annotations: map[string]string{"node.machine.sapcloud.io/not-managed-by-mcm": "1"},
+							},
+						},
+					},
+				}
+			)
+
+			Expect(CheckNodesScalingDown(machineList, nodeList, 2, 1)).To(MatchError(ContainSubstring("is waiting to be completely drained from pods")))
+		})
+	})
 })
 
 func newDeployment(namespace, name, role string, healthy bool) *appsv1.Deployment {
@@ -766,6 +1180,32 @@ func newEtcd(namespace, name, role string, healthy bool, lastError *string) *dru
 	return etcd
 }
 
+func newNode(name string, healthy bool, labels labels.Set, annotations map[string]string, kubeletVersion string) corev1.Node {
+	node := corev1.Node{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:        name,
+			Labels:      labels,
+			Annotations: annotations,
+		},
+		Status: corev1.NodeStatus{
+			NodeInfo: corev1.NodeSystemInfo{
+				KubeletVersion: kubeletVersion,
+			},
+		},
+	}
+
+	if healthy {
+		node.Status.Conditions = []corev1.NodeCondition{
+			{
+				Type:   corev1.NodeReady,
+				Status: corev1.ConditionTrue,
+			},
+		}
+	}
+
+	return node
+}
+
 func roleLabels(role string) map[string]string {
 	return map[string]string{v1beta1constants.GardenRole: role}
 }
@@ -788,4 +1228,8 @@ func roleOf(obj metav1.Object) string {
 
 func beConditionWithStatusAndCodes(status gardencorev1beta1.ConditionStatus, codes ...gardencorev1beta1.ErrorCode) types.GomegaMatcher {
 	return And(WithStatus(status), WithCodes(codes...))
+}
+
+func beConditionWithStatusAndMsg(status gardencorev1beta1.ConditionStatus, reason, message string) types.GomegaMatcher {
+	return And(WithStatus(status), WithReason(reason), WithMessage(message))
 }
