+++ skaffold-operator.yaml
@@ -65,7 +65,6 @@ build:
         - pkg/component/clusteridentity
         - pkg/component/coredns
         - pkg/component/coredns/constants
-        - pkg/component/dependencywatchdog
         - pkg/component/etcd
         - pkg/component/etcd/constants
         - pkg/component/etcdcopybackupstask
@@ -74,10 +73,8 @@ build:
         - pkg/component/extensions/dnsrecord
         - pkg/component/extensions/extension
         - pkg/component/extensions/infrastructure
-        - pkg/component/extensions/network
         - pkg/component/extensions/operatingsystemconfig
         - pkg/component/extensions/operatingsystemconfig/downloader
-        - pkg/component/extensions/operatingsystemconfig/executor
         - pkg/component/extensions/operatingsystemconfig/original
         - pkg/component/extensions/operatingsystemconfig/original/components
         - pkg/component/extensions/operatingsystemconfig/original/components/containerd
@@ -108,24 +105,18 @@ build:
         - pkg/component/kubescheduler
         - pkg/component/kubestatemetrics
         - pkg/component/logging
-        - pkg/component/logging/eventlogger
         - pkg/component/logging/fluentoperator
         - pkg/component/logging/fluentoperator/customresources
-        - pkg/component/logging/kuberbacproxy
         - pkg/component/logging/vali
         - pkg/component/machinecontrollermanager
-        - pkg/component/metricsserver
         - pkg/component/monitoring
-        - pkg/component/namespaces
         - pkg/component/nginxingress
         - pkg/component/nodeexporter
         - pkg/component/nodelocaldns
         - pkg/component/nodelocaldns/constants
-        - pkg/component/nodeproblemdetector
         - pkg/component/plutono
         - pkg/component/resourcemanager
         - pkg/component/resourcemanager/constants
-        - pkg/component/shootsystem
         - pkg/component/shared
         - pkg/component/vpa
         - pkg/component/vpa/constants
@@ -148,12 +139,6 @@ build:
         - pkg/gardenlet/apis/config/v1alpha1
         - pkg/healthz
         - pkg/logger
-        - pkg/operation
-        - pkg/operation/botanist
-        - pkg/operation/common
-        - pkg/operation/garden
-        - pkg/operation/seed
-        - pkg/operation/shoot
         - pkg/operator/apis/config
         - pkg/operator/apis/config/v1alpha1
         - pkg/operator/apis/config/validation
@@ -191,7 +176,6 @@ build:
         - pkg/utils/gardener/tokenrequest
         - pkg/utils/imagevector
         - pkg/utils/kubernetes
-        - pkg/utils/kubernetes/client
         - pkg/utils/kubernetes/health
         - pkg/utils/kubernetes/health/checker
         - pkg/utils/kubernetes/unstructured
@@ -200,7 +184,6 @@ build:
         - pkg/utils/retry
         - pkg/utils/secrets
         - pkg/utils/secrets/manager
-        - pkg/utils/time
         - pkg/utils/timewindow
         - pkg/utils/validation/admissionplugins
         - pkg/utils/validation/cidr
